fix: fix overlap of rotated inside labels in pie chart
