+++ packages/vrender-components/src/axis/overlap/auto-hide.ts
@@ -2,7 +2,7 @@
  * @description 自动隐藏
  */
 
-import { createRect, type IText } from '@visactor/vrender-core';
+import { type IText } from '@visactor/vrender-core';
 // eslint-disable-next-line no-duplicate-imports
 import { isEmpty, isFunction, last } from '@visactor/vutils';
 import type { CustomMethod } from '../type';
